`fn Rav1dPicAllocator::{de,}alloc_picture_data`: Move to `Rav1dPictureData` `mod` for private field access.
